+++ core/src/Extensions.kt
@@ -8,6 +8,7 @@ import kotlin.reflect.KType
 fun String?.trimToNull() = this?.trim()?.takeIf { it.isNotEmpty() }
 
 fun OutputStream.write(s: String) = write(s.toByteArray())
+fun OutputStream.writeln() { write(13); write(10) }
 
 @Suppress("UNCHECKED_CAST")
 fun <K, V> notNullValues(vararg pairs: Pair<K, V?>?) = pairs.filter { it?.second != null } as List<Pair<K, V>>