+++ server/test/klite/MultipartRendererTest.kt
@@ -12,14 +12,23 @@ class MultipartRendererTest {
     expect(renderer.fullContentType).toEqual(MimeTypes.formData + "; boundary=MyBoundary")
 
     val output = ByteArrayOutputStream()
-    renderer.render(output, mapOf("name1" to "value1", "file2" to FileUpload("a.txt", "text/plain", "Content of a.txt.".byteInputStream())))
+    renderer.render(output, mapOf(
+      "name1" to "value1",
+      "name2" to "Hello".toByteArray(),
+      "my-file" to FileUpload("a.txt", "text/plain", "Content of a.txt.".byteInputStream()),
+      "null" to null
+    ))
     expect(output.toString()).toEqual("""
       --MyBoundary
       Content-Disposition: form-data; name="name1"
 
       value1
       --MyBoundary
-      Content-Disposition: form-data; name="file2"; filename="a.txt"
+      Content-Disposition: form-data; name="name2"
+
+      Hello
+      --MyBoundary
+      Content-Disposition: form-data; name="my-file"; filename="a.txt"
       Content-Type: text/plain; charset=UTF-8
 
       Content of a.txt.